NU-2436 rename displayTestResults to testingActions
