+++ designer/client/src/actions/reduxTypes.ts
@@ -9,7 +9,6 @@ import type { ScenariosActions } from "../reducers/scenarios";
 import type { ActionTypes } from "./actionTypes";
 import type { AssistantActions } from "./assistantActions";
 import type { CountsActions } from "./nk/displayProcessCounts";
-import type { TestsActions } from "./nk/displayTestResults";
 import type { PropertiesActions } from "./nk/editProperties";
 import type { LiveDataActions } from "./nk/liveData";
 import type { NodeActions } from "./nk/node";
@@ -19,6 +18,7 @@ import type { ScenarioActions, UpdateTestCapabilitiesAction } from "./nk/process
 import type { GetScenarioActivitiesAction, UpdateScenarioActivitiesAction } from "./nk/scenarioActivities";
 import type { SelectionActions } from "./nk/selection";
 import type { TestCasesActions } from "./nk/testCasesActions";
+import type { TestsActions } from "./nk/testingActions";
 import type { ToolbarActions } from "./nk/toolbars";
 import type { ToolWindowActions } from "./nk/toolWindow";
 import type { UiActions } from "./nk/ui/uiActions";
